pyproject.toml
[project]
name = "slashrun"
version = "0.1.0"
description = "AI-powered simulation platform for research and wargaming"
requires-python = ">=3.12"

dependencies = [
    "aiohttp>=3.12.15",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "cryptography>=45.0.7",
    "email-validator>=2.1.0",
    "fastapi>=0.116.1",
    "fredapi>=0.5.2",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.19",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langsmith>=0.4.25",
    "newsapi-python>=0.2.7",
    "pandas>=2.3.2",
    "passlib>=1.7.4",
    "pgvector>=0.4.1",
    "psycopg>=3.2.10",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.6",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
    "wbgapi>=1.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "aiosqlite>=0.21.0",
    "bcrypt>=4.3.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--cov=backend/app",
    "--cov-report=term-missing",
    "--cov-report=html:coverage_html"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
